idefics-docker/requirements.txt
flask==2.2.3
flask-cors==3.0.10
text-generation==0.6.0
requests==2.28.2
python-dotenv==1.0.0
werkzeug==2.2.3